feat: update posts module

+++ internal/models/post.go
@@ -4,6 +4,6 @@ type Post struct {
 	ID        uint64 `json:"id"`
 	Content   string `json:"content"`
 	Title     string `json:"title"`
-	UserID    int    `json:"user_id"`
+	UserID    uint64 `json:"user_id"`
 	CreatedAt string `json:"created_at"`
 }
